requirements.txt
Flask
transformers
Pillow
torch 
torchvision 
torchaudio